Create MultSplitPanels.java
As java has just a two Component split system the JSplitPane, I created a new JSplitPane that can be placed as many components as wanted. I am still developing it, but I want to share and maybe I get some help and ideias to organize screens in Java. Sorry for using portuguese in some methods, but as I said I didnt finished yet.